# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CraveDash App API",
    "version": "1.0.0",
    "description": "API documentation for the Fast Food Order and Delivery Management System."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "HTTP local server"
    },
    {
      "url": "https://localhost:3443/api",
      "description": "HTTPS local server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Invalid credentials"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
      "StkPushRequest": {
        "type": "object",
        "required": [
          "orderId",
          "phone"
        ],
        "properties": {
          "orderId": {
            "type": "integer",
            "example": 101
          },
          "phone": {
            "type": "string",
            "example": "254712345678"
          }
        }
      },
      "StkCallbackRequest": {
        "type": "object",
        "required": [
          "orderId",
          "resultCode",
          "resultDesc"
        ],
        "properties": {
          "orderId": {
            "type": "integer",
            "example": 101
          },
          "resultCode": {
            "type": "integer",
            "example": 0
          },
          "resultDesc": {
            "type": "string",
            "example": "Success"
          },
          "mpesaReceiptNumber": {
            "type": "string",
            "example": "QWE1234567"
          }
        }
      },
...
      "StkPushData": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "checkoutRequestId": {
            "type": "string",
            "example": "ws_CO_101_1709541455000_b0c18aa793b09fd4"
          },
          "merchantRequestId": {
            "type": "string",
            "example": "MR-1709541455000"
          },
          "responseCode": {
            "type": "string",
            "example": "0"
          },
          "responseDescription": {
            "type": "string",
            "example": "Success. Request accepted for processing"
          },
          "customerMessage": {
            "type": "string",
            "example": "Please enter your M-Pesa PIN on your phone"
          },
          "orderId": {
            "type": "integer",
            "example": 101
          },
          "amount": {
            "type": "number",
            "format": "float",
            "example": 900
          }
        }
      },
      "StkCallbackData": {
        "type": "object",
        "properties": {
          "ResultCode": {
            "type": "integer",
            "example": 0
          },
          "ResultDesc": {
            "type": "string",
            "example": "Success"
          },
          "MpesaReceiptNumber": {
            "type": "string",
            "example": "SIM17095415"
          }
        },
        "required": [
          "ResultCode",
          "ResultDesc"
        ]
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "Too many requests",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/payments/stk-push": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Initiate STK push (authenticated)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StkPushRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "STK push initiated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "STK push initiated. Please check your phone for the payment prompt."
                    },
                    "data": {
                      "$ref": "#/components/schemas/StkPushData"
                    }
                  }
                },
                "examples": {
                  "success": {
                    "value": {
                      "success": true,
                      "message": "STK push initiated. Please check your phone for the payment prompt.",
                      "data": {
                        "success": true,
                        "checkoutRequestId": "ws_CO_101_1709541455000_b0c18aa793b09fd4",
                        "merchantRequestId": "MR-1709541455000",
                        "responseCode": "0",
                        "responseDescription": "Success. Request accepted for processing",
                        "customerMessage": "Please enter your M-Pesa PIN on your phone",
                        "orderId": 101,
                        "amount": 900
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
    "/payments/stk-callback": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Simulated STK callback",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StkCallbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Callback processed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/StkCallbackData"
                    }
                  }
                },
                "examples": {
                  "success": {
                    "value": {
                      "success": true,
                      "message": "Success",
                      "data": {
                        "ResultCode": 0,
                        "ResultDesc": "Success",
                        "MpesaReceiptNumber": "SIM17095415"
                      }
                    }
                  },
                  "failedPayment": {
                    "value": {
                      "success": true,
                      "message": "Success",
                      "data": {
                        "ResultCode": 1032,
                        "ResultDesc": "Request cancelled by user"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
...
  }
}